# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for authenticated Linux users to manage Containerlab deployments, topology files, node access, captures, and host networking tools.\n\nMost endpoints under /api/v1 require a JWT from POST /login. Send it as: Authorization: Bearer \u003ctoken\u003e.",
        "title": "Containerlab API Server",
        "contact": {
            "name": "Containerlab API Server maintainers",
            "url": "https://github.com/srl-labs/clab-api-server/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/capabilities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the API contract version, server version, runtime, and optional protocol features supported by this build.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "Get API capabilities",
                "responses": {
                    "200": {
                        "description": "API capabilities",
                        "schema": {
                            "$ref": "#/definitions/models.CapabilitiesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/labs/{labName}/nodes/{nodeName}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns logs for a lab node.\n\n**Notes**\n- When `follow=true`, the response streams as NDJSON (one JSON object per line) until the client disconnects or the 30-minute timeout.",
                "produces": [
                    "application/json",
                    "application/x-ndjson"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Get node logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the lab",
                        "name": "labName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Full name of the container (node)",
                        "name": "nodeName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "all",
                        "description": "Number of lines to show from the end of logs (default all). Use an integer or 'all'.",
                        "name": "tail",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Follow log output (stream logs as NDJSON). Note: In Swagger UI, streaming may not display correctly.",
                        "name": "follow",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Container logs (follow=false). When follow=true, response is NDJSON stream of LogLine objects.",
                        "schema": {
                            "$ref": "#/definitions/models.LogsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input (lab name, node filter, etc.)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden (not owner of the lab)",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Lab or node not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/labs/{labName}/topology/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Streams topology YAML/annotations change events for a single lab topology document pair as NDJSON.",
                "produces": [
                    "application/x-ndjson"
                ],
                "tags": [
                    "Labs"
                ],
                "summary": "Stream topology document events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lab name",
                        "name": "labName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative topology YAML or annotations path inside lab directory",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Topology document event stream",
                        "schema": {
                            "$ref": "#/definitions/models.TopologyDocEventResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid path",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/session": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the stable API identity, roles, and bearer-token lifetime for the current caller.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "Get authenticated session",
                "responses": {
                    "200": {
                        "description": "Authenticated session",
                        "schema": {
                            "$ref": "#/definitions/models.SessionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CapabilitiesResponse": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string",
                    "example": "v1"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "runtime": {
                    "type": "string",
                    "example": "docker"
                },
                "serverVersion": {
                    "type": "string",
                    "example": "v0.5.0"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid credentials or user not in allowed group"
                }
            }
        },
...
        "models.LogsResponse": {
            "type": "object",
            "properties": {
                "containerName": {
                    "type": "string"
                },
                "logs": {
                    "type": "string"
                }
            }
        },
...
        "models.SessionResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "issuedAt": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.TopologyDocEventResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "change"
                },
                "documentKind": {
                    "type": "string",
                    "example": "annotations"
                },
                "labName": {
                    "type": "string",
                    "example": "mylab"
                },
                "path": {
                    "type": "string",
                    "example": "mylab.clab.yml.annotations.json"
                },
                "revision": {
                    "type": "string",
                    "example": "yaml=123-456;annotations=78-910"
                },
                "type": {
                    "type": "string",
                    "example": "topology-doc"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}